+++ NEWS.md
@@ -1,3 +1,5 @@
+# downlit (development version)
+
 # downlit 0.4.1
 
 ## Syntax highlighting