+++ NEWS.md
@@ -3,6 +3,8 @@
 * `highlight()` no longer errors if a package imputed to have been attached
   isn't installed.
 
+* Correctly link `requireNamespace(MASS)` (#151).
+
 # downlit 0.4.1
 
 ## Syntax highlighting