+++ NEWS.md
@@ -1,4 +1,4 @@
-# downlit (development version)
+# downlit 0.4.2
 
 * `highlight()` no longer errors if a package imputed to have been attached
   isn't installed.